refactor(element quiz): separate data set
