+++ openday_scavenger/puzzles/element/views.py
@@ -69,7 +69,7 @@ async def index(
         context={
             "puzzle": puzzle_name,
             "visitor": visitor.uid,
-            "location": suffix.title(),
+            "location": suffix.upper(),
             "elements": elements,
             "element_lookup": element_lookup,
             "get_category_style": get_category_style,
